Switched path management over to use Bundler
