+++ lib/wrest/core_ext/hash.rb
@@ -1,4 +1,4 @@
-require "#{Wrest::Root}/wrest/core_ext/hash/conversions"
+require "wrest/core_ext/hash/conversions"
 
 class Hash #:nodoc:
   include Wrest::CoreExt::Hash::Conversions
