+++ lib/wrest/core_ext/string.rb
@@ -1,4 +1,4 @@
-require "#{Wrest::Root}/wrest/core_ext/string/conversions"
+require "wrest/core_ext/string/conversions"
 
 class String #:nodoc:
   include Wrest::CoreExt::String::Conversions unless Wrest.const_defined?('NoStringExtensions')
