+++ lib/wrest/http_shared.rb
@@ -14,9 +14,9 @@ module Wrest
   end
 end
 
-require "#{Wrest::Root}/wrest/http_shared/headers"
-require "#{Wrest::Root}/wrest/http_shared/standard_headers"
-require "#{Wrest::Root}/wrest/http_shared/standard_tokens"
+require "wrest/http_shared/headers"
+require "wrest/http_shared/standard_headers"
+require "wrest/http_shared/standard_tokens"
 
 # Set up a shorter convenience API for constants
 Wrest::H = Wrest::HttpShared::StandardHeaders
