+++ lib/wrest/test.rb
@@ -1,1 +1,1 @@
-require "#{Wrest::Root}/wrest/test/request_patches"
+require "wrest/test/request_patches"
