+++ spec/wrest/callback_spec.rb
@@ -1,5 +1,4 @@
 require "spec_helper"
-require "#{Wrest::Root}/wrest"
 
 module Wrest
   describe Callback do
