+++ spec/wrest/components/translators/xml_spec.rb
@@ -27,7 +27,7 @@ module Wrest::Components::Translators
       Xml.deserialise(http_response)
     end
     
-    xml_backends.each do |e| 
+    Helpers.xml_backends.each do |e| 
       it "should be able to pull out desired elements from an xml response based on xpath and return an array of matching nodes" do
         ActiveSupport::XmlMini.backend = e
